+++ src/test/resources/inspection/knownConnection.php
@@ -1,0 +1,3 @@
+<?php
+\Hyperf\DbConnection\Db::connection('goods');
+\Hyperf\DbConnection\Schema::connection('default');