+++ src/test/resources/inspection/knownDbTable.php
@@ -1,0 +1,1 @@
+<?php \Hyperf\DbConnection\Db::table('users');