+++ src/test/resources/inspection/unknownDbTable.php
@@ -1,0 +1,1 @@
+<?php \Hyperf\DbConnection\Db::table('<warning descr="Unknown table or view">wrongTable</warning>');