+++ src/test/resources/inspection/unknownModelConnectionProperty.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App {
+    class UserModel extends \Hyperf\Database\Model\Model
+    {
+        protected ?string $connection = '<warning descr="Unknown database connection">missing</warning>';
+    }
+}