+++ src/test/resources/inspection/unknownModelTableProperty.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App {
+    class GoodsModel extends \Hyperf\Database\Model\Model
+    {
+        protected $table = '<warning descr="Unknown table or view">goods</warning>';
+    }
+}