+++ src/test/resources/model/modelTablePropertyOtherConnectionReference.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App {
+    class UserModel extends \Hyperf\Database\Model\Model
+    {
+        protected ?string $connection = 'goods';
+
+        protected $table = 'users';
+    }
+}